+++ Source/OS9exec_core/utilstuff.c
@@ -1540,7 +1540,7 @@ os9err Flush_Dir( ushort cpid, ushort* pathP, const char* nmS )
 {
   os9err          err;
   os9direntry_typ d;
-  ulong           dir_size;
+  uint32_t        dir_size;
   syspath_typ*    spP;
   char            fullName[OS9PATHLEN];
   int             oLen;
